requirements.txt
pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=2.0.0
customtkinter>=5.2.0
cairosvg